Rename backend to `word-history-explorer-backend`
This is the first of two commits trying to clean up some confusing
naming resulting from the rust-wasm templates. When following the
official "Rust and WebAssembly" tutorial one ends up with a directory
structure that contains two npm packages:

* Running `wasm-pack build` generates a subdirectory `pkg` that contains
  a `package.json` file with a "name" attribute taken from `Cargo.toml`.

* Running `npm init wasm-app www` generates a subdirectory `www` that
  contains a `package.json` file with a "name" attribute set to
  "create-wasm-app", which is not very descriptive.

This is confusing because the package in the `www` subdirectory (with
the generic name "create-wasm-app") is really the "main" package, while
the package in the `pkg` subdirectory (with the more descriptive name
taken from `Cargo.toml`) is only a dependency of the main package. When
one eventually builds the website by running `npm run build` or
`npm run start` from the `www` subdirectory, the package `www` will be
bundled into or served as a website, and the package from the `pkg`
subdirectory will only be one of the dependencies of the bundle.

To make this more clear: a typical web app will not be written purely
in Rust. More likely, the Rust+WebAssembly part will perform some
computationally heavy tasks while the UI may be more conveniently be
implemented in JavaScript. In the structure set up by the templates from
the "Rust and WebAssembly" tutorial, one would have to write this
JavaScript UI code inside a module called "create-wasm-app", which seems
odd.

This commit changes the name of the package in the subdirectory `pkg`
to `word-history-explorer-backend`, adding the `-backend` suffix. The
next commit will change the name of the package in the `www`
subdirectory to `word-history-explorer`, thus giving it a descriptive
name.