Implement lossless compression with ANS
ANS is short for "Asymmetric Numeral System". It's an entropy coding
that has essentially the same compression rate as arithmetic coding but
runs much faster. ANS treats the compressed file as a stack. The
pushes compressed symbols (with fractional bit lengths) on the stack in
reverse order so that the decoder can pop them off the stack in the
expected order.

Naively, pushing data with fractional bit length on or popping it off
a stack would change the (fractional) alignment of the entire data on
the stack. This would mean that each push and pop operation would
require an expensive realignment of the entire data on the stack,
rendering CPU costs for encoding and decoding to O(length^2). To avoid
this, ANS uses a second, very short "work" stack on which it accumulates
information with fractional bit lengths. When pushing more information
would overflow the work stack, ANS pops off a fixed size chunk of
integer bit length (in practice: a multiple of bytes) from the work
stack and pushes it on the main stack. Since the chunk is integer sized,
pushing it on the main stack requires no realignment of existing data on
the main stack. The Decoder operates in reverse order. At the end, the
encoder pads the work stack and flushes it on the main stack. The
decoder operates in reverse order. It initializes its work stack from
the top of the stack and then continuously drains it. When the work
work stack falls below a critical filling level, defined by details of
the protocol, then the decoder pops a fixed size chunk off the main
stack and pushes it on the work stack.

This commit implements two variants of ANS. Both variants use a work
stack of 32 bits, i.e., a simple `u32`. One variant transfers data
between work and main stack in chunks of 8 bits, and the other transfers
data in chunks of 16 bits. The latter variant (with 16 bit chunks) seems
preferable since both variants reach essentially indistinguishable
compression performance but the large chunk size speeds up decompression
by about 5 to 10%.

The encoders of both variants currently use some unsafe code to avoid
bounds checks in a hot loop. This is mainly to simplify benchmarking.
In an upcoming commit, I will re-introducing bounds checks and measure
their effect on performance. The impact of bounds checks should decrease
with increasing chunk size, so if anything, it should further increase
the benefits of 16 bit sized chunks compared to 8 bit sized chunks.