Implement memory mapping for embedding file
Add a module that provides simple and checked access to data in a custom
binary file format for compressed dynamic word embeddings.

The file format is described in `file-format.html`. The implementation
allows structured access to the contents of the file with minimal data
copying. It essentially just holds a shared reference to the entire file
as a `[u32]`, and it returns views into the data with convenient types.
The module also facilitates the construction of decoders for compressed
chunks of the file.