Load initial configuration from URL
This commit changes the format of constructed URLs slightly: instead of
appending the configuration directly to the path, we now separate the
configuration from the path with a pound sign "#". For example, this
could result in the following URL:

    http://localhost:8080/#hey_@_there

This has the advantage that, when a web browser sees such an address, it
will only request the base URL (here: "http://localhost:8080/") from the
server, so it will load the correct web app.

In addition, the web app now reads out the configuration from the
address once it's loaded. Thus, we can now share URLs for specific
configurations of the app with others (of course, this will only be
useful once we host the app on a publicly visible URL).