Outsource entropy coding to `constriction`
This replaces our internal entropy coder with the `AnsCoder` provided by
the `constriction` crate, allowing us to remove the entire
`compression.rs` file with 1278 lines of code.

This is a port with minimal changes, which actually leads to somewhat
convoluted code. Had we used an external library from the start we would
probably have ended up with simpler code.