Fix bug in CLI
Subcommand `create` of the CLI did not quite conform to the
documentation in `file-format.html` as it initialized `state` for
compression with zero rather than `1^16`. It turns out this didn't make
a difference on embedding vectors we currently use in the web app
(i.e., re-encoding them with the fixed implementation resulted in a
binary identical `.dwe` file) but it might make a difference for
different embeddings, and fixing it also makes the implementation easier
to reason about.