Fix bug in CLI
The subcommand `create` of the CLI app did not quite conform to the
documentation in `file-format.html` as it initialized `state` for
compression with zero rather than `1^16`. This probably would have led
to decoding errors in the very last embedding vector if the compressed
embedding vector section needed padding to a multiple of 4 bytes.
However, it doesn't affect the `.dwe` file currently used by the web app
since, as it turns out, that file was created before the bug that's
fixed by this commit was introduced.